requirements.txt
APScheduler==3.10.0
asgiref==3.4.1
Django==3.2.6
django-apscheduler==0.6.2
django-ckeditor==6.1.0
django-crispy-forms==1.12.0
django-js-asset==1.2.2
django-phonenumber-field==5.2.0
django-slick-reporting==0.6.6
gunicorn==20.1.0
phonenumbers==8.12.29
Pillow==8.3.2
psycopg2==2.9.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
simplejson==3.18.3
six==1.16.0
sqlparse==0.4.1
tzdata==2022.7
tzlocal==4.2
whitenoise==5.3.0
